+++ src/main/java/org/prgrms/spring_week1/services/io/Input.java
@@ -1,5 +1,6 @@
 package org.prgrms.spring_week1.services.io;
 
 public interface Input {
-    void input();
+    String input();
+    String inputInLine();
 }
