+++ src/main/java/org/prgrms/spring_week1/services/io/Output.java
@@ -2,5 +2,7 @@ package org.prgrms.spring_week1.services.io;
 
 public interface Output {
     void output(String str);
-    void menu();
+    void mainMenu();
+    void voucherCreate();
+    void wrongInput();
 }
